--- src/main/java/com/woowacourse/pelotonbackend/domain/RaceCategory.java
@@ -1,5 +1,0 @@
-package com.woowacourse.pelotonbackend.domain;
-
-public enum RaceCategory {
-    TIME
-}
