--- src/main/java/com/woowacourse/pelotonbackend/domain/Role.java
@@ -1,5 +1,0 @@
-package com.woowacourse.pelotonbackend.domain;
-
-public enum Role {
-    MEMBER, ADMIN
-}
