+++ src/main/java/com/woowacourse/pelotonbackend/member/Role.java
@@ -1,0 +1,5 @@
+package com.woowacourse.pelotonbackend.member;
+
+public enum Role {
+    MEMBER, ADMIN
+}
