+++ src/main/java/com/woowacourse/pelotonbackend/mission/Mission.java
@@ -1,8 +1,7 @@
-package com.woowacourse.pelotonbackend.domain;
+package com.woowacourse.pelotonbackend.mission;
 
 import java.time.LocalDateTime;
 
-import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 
 import org.springframework.data.annotation.CreatedDate;
@@ -11,6 +10,7 @@ import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jdbc.core.mapping.AggregateReference;
 import org.springframework.data.relational.core.mapping.Embedded;
 
+import com.woowacourse.pelotonbackend.race.Race;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -22,24 +22,18 @@ import lombok.With;
 @Builder
 @EqualsAndHashCode(of = "id")
 @Getter
-public class MissionCertification {
+public class Mission {
     @Id @With(value = AccessLevel.PACKAGE)
     private final Long id;
 
-    @NotNull
-    private final MissionCertificationStatus status;
-
-    @NotBlank
-    private final String description;
+    @Embedded.Empty
+    private final TimeDuration missionDuration;
 
     @Embedded.Empty
-    private final ImageUrl image;
+    private final MissionInstruction missionInstruction;
 
     @NotNull
-    private final AggregateReference<Rider, @NotNull Long> riderId;
-
-    @NotNull
-    private final AggregateReference<Mission, @NotNull Long> missionId;
+    private final AggregateReference<Race, Long> raceId;
 
     @CreatedDate
     private LocalDateTime createdAt;
