+++ src/main/java/com/woowacourse/pelotonbackend/missioncertification/MissionCertification.java
@@ -1,16 +1,19 @@
-package com.woowacourse.pelotonbackend.domain;
+package com.woowacourse.pelotonbackend.missioncertification;
 
 import java.time.LocalDateTime;
 
-import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jdbc.core.mapping.AggregateReference;
 import org.springframework.data.relational.core.mapping.Embedded;
 
+import com.woowacourse.pelotonbackend.vo.ImageUrl;
+import com.woowacourse.pelotonbackend.rider.Rider;
+import com.woowacourse.pelotonbackend.mission.Mission;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -22,22 +25,24 @@ import lombok.With;
 @Builder
 @EqualsAndHashCode(of = "id")
 @Getter
-public class Member {
+public class MissionCertification {
     @Id @With(value = AccessLevel.PACKAGE)
     private final Long id;
 
-    @NotBlank
-    private final String name;
+    @NotNull
+    private final MissionCertificationStatus status;
 
     @NotBlank
-    @Email
-    private final String email;
+    private final String description;
 
     @Embedded.Empty
-    private final Cash cash;
+    private final ImageUrl image;
 
     @NotNull
-    private final Role role;
+    private final AggregateReference<Rider, @NotNull Long> riderId;
+
+    @NotNull
+    private final AggregateReference<Mission, @NotNull Long> missionId;
 
     @CreatedDate
     private LocalDateTime createdAt;
