+++ src/main/java/com/woowacourse/pelotonbackend/missioncertificationreport/MissionCertificationReport.java
@@ -1,4 +1,4 @@
-package com.woowacourse.pelotonbackend.domain;
+package com.woowacourse.pelotonbackend.missioncertificationreport;
 
 import java.time.LocalDateTime;
 
@@ -8,8 +8,10 @@ import javax.validation.constraints.NotNull;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.relational.core.mapping.Embedded;
+import org.springframework.data.jdbc.core.mapping.AggregateReference;
 
+import com.woowacourse.pelotonbackend.member.Member;
+import com.woowacourse.pelotonbackend.missioncertification.MissionCertification;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -21,30 +23,21 @@ import lombok.With;
 @Builder
 @EqualsAndHashCode(of = "id")
 @Getter
-public class Race {
+public class MissionCertificationReport {
     @Id @With(value = AccessLevel.PACKAGE)
     private final Long id;
 
-    @NotBlank
-    private final String title;
+    @NotNull
+    private final ReportType reportType;
 
     @NotBlank
     private final String description;
 
-    @Embedded.Empty
-    private final ImageUrl thumbnail;
-
-    @Embedded.Empty
-    private final ImageUrl certificationExample;
-
-    @Embedded.Empty
-    private final DateDuration raceDuration;
+    @NotNull
+    private final AggregateReference<MissionCertification, @NotNull Long> missionCertificationId;
 
     @NotNull
-    private final RaceCategory category;
-
-    @Embedded.Empty
-    private final Cash entranceFee;
+    private final AggregateReference<Member, @NotNull Long> memberId;
 
     @CreatedDate
     private LocalDateTime createdAt;
