+++ src/main/java/com/woowacourse/pelotonbackend/race/DateDuration.java
@@ -1,4 +1,4 @@
-package com.woowacourse.pelotonbackend.domain;
+package com.woowacourse.pelotonbackend.race;
 
 import java.time.LocalDate;
