+++ src/main/java/com/woowacourse/pelotonbackend/race/Race.java
@@ -1,14 +1,17 @@
-package com.woowacourse.pelotonbackend.domain;
+package com.woowacourse.pelotonbackend.race;
 
 import java.time.LocalDateTime;
 
+import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jdbc.core.mapping.AggregateReference;
+import org.springframework.data.relational.core.mapping.Embedded;
 
+import com.woowacourse.pelotonbackend.vo.Cash;
+import com.woowacourse.pelotonbackend.vo.ImageUrl;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -20,15 +23,30 @@ import lombok.With;
 @Builder
 @EqualsAndHashCode(of = "id")
 @Getter
-public class Rider {
+public class Race {
     @Id @With(value = AccessLevel.PACKAGE)
     private final Long id;
 
-    @NotNull
-    private final AggregateReference<Member, @NotNull Long> memberId;
+    @NotBlank
+    private final String title;
+
+    @NotBlank
+    private final String description;
+
+    @Embedded.Empty
+    private final ImageUrl thumbnail;
+
+    @Embedded.Empty
+    private final ImageUrl certificationExample;
+
+    @Embedded.Empty
+    private final DateDuration raceDuration;
 
     @NotNull
-    private final AggregateReference<Race, @NotNull Long> raceId;
+    private final RaceCategory category;
+
+    @Embedded.Empty
+    private final Cash entranceFee;
 
     @CreatedDate
     private LocalDateTime createdAt;
