+++ src/main/java/com/woowacourse/pelotonbackend/race/RaceCategory.java
@@ -1,0 +1,5 @@
+package com.woowacourse.pelotonbackend.race;
+
+public enum RaceCategory {
+    TIME
+}
