+++ src/main/java/com/woowacourse/pelotonbackend/rider/Rider.java
@@ -1,16 +1,16 @@
-package com.woowacourse.pelotonbackend.domain;
+package com.woowacourse.pelotonbackend.rider;
 
 import java.time.LocalDateTime;
 
-import javax.validation.constraints.FutureOrPresent;
 import javax.validation.constraints.NotNull;
 
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jdbc.core.mapping.AggregateReference;
-import org.springframework.data.relational.core.mapping.Embedded;
 
+import com.woowacourse.pelotonbackend.member.Member;
+import com.woowacourse.pelotonbackend.race.Race;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -22,18 +22,15 @@ import lombok.With;
 @Builder
 @EqualsAndHashCode(of = "id")
 @Getter
-public class Mission {
+public class Rider {
     @Id @With(value = AccessLevel.PACKAGE)
     private final Long id;
 
-    @Embedded.Empty
-    private final TimeDuration missionDuration;
-
-    @Embedded.Empty
-    private final MissionInstruction missionInstruction;
+    @NotNull
+    private final AggregateReference<Member, @NotNull Long> memberId;
 
     @NotNull
-    private final AggregateReference<Race, Long> raceId;
+    private final AggregateReference<Race, @NotNull Long> raceId;
 
     @CreatedDate
     private LocalDateTime createdAt;
