+++ src/main/java/com/woowacourse/pelotonbackend/vo/Cash.java
@@ -1,4 +1,4 @@
-package com.woowacourse.pelotonbackend.domain;
+package com.woowacourse.pelotonbackend.vo;
 
 import java.math.BigDecimal;
