+++ src/main/java/com/woowacourse/pelotonbackend/vo/ImageUrl.java
@@ -1,4 +1,4 @@
-package com.woowacourse.pelotonbackend.domain;
+package com.woowacourse.pelotonbackend.vo;
 
 import javax.validation.constraints.NotBlank;
